requirements.txt
streamlit==1.31.0
anthropic>=0.18.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.9
fpdf==1.7.2
Pillow==10.2.0
python-dotenv==1.0.1